Collision, Animations

+++ src/main/java/moe/cameo/units/Unit.java
@@ -1,9 +1,9 @@
 package moe.cameo.units;
 
 public class Unit {
-    private float x;
-    private float y;
+    private int x;
+    private int y;
 
-    public float getX() { return x; }
-    public float getY() { return y; }
+    public int getX() { return x; }
+    public int getY() { return y; }
 }
